requirements.txt
numpy==1.26.4
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.11.7
yt-dlp==2025.6.9
openai-whisper==20240930
requests==2.32.4
opencv-python==4.11.0.86
python-multipart==0.0.19